Move WalletBadge component to separate file and update layout import
